refactor(projects): extract EquipmentAddForm (Phase 4b)
Split AddEquipmentDialog into a thin <Dialog> wrapper (22 lines) and an
embeddable EquipmentAddForm body so the form can also be hosted inside
the upcoming UnifiedAddDialog without duplicating the model/asset-tag
picker, availability check, and overbook/duplicate handling.

Pure refactor:
- No prop changes on AddEquipmentDialog — existing callers (toolbar
  Add Equipment button, group-kebab onAddEquipment) keep working
  unchanged.
- Form body identical to previous implementation: mode toggle,
  ComboboxPicker model search, asset-tag lookup, availability /
  overbook / duplicate panels, quantity / unit-price / discount /
  category / notes / isOptional, Add to Project submit.
- Verified manually: dialog opens, both modes render, mode toggle
  preserves disabled-quantity behavior in asset-tag mode, console
  clean (only the pre-existing <table>/<div> nesting warning from
  EquipmentTab itself).

Mirrors the Phase 4a extraction shape (CustomItemAddForm) so the
UnifiedAddDialog shell can mount both bodies through a single segmented
switcher in Phase 4d.
